LLM/requirements.txt
streamlit==1.29.0
openai>=1.30.0
pandas==2.1.4
python-dotenv==1.0.0
plotly==5.18.0
pyyaml==6.0.1
